+++ CHANGELOG.md
@@ -6,9 +6,10 @@
 - value count for each tab
 - number of available value types in entry picker (per category and in total)
 - visual marker for timezone-dependant values
-- "What's New" under Settings (references CHANGELOG.md on GitHub main branch, = this file)
 - option to expand the date value type with ISO calendar week and day of the year
-- value type: Modified Julian Date 2000
+- value type: Modified Julian Date 2000 (MJD2000)
+- value type: Local Mean Sidereal Time (LMST)
+- "What's New" under Settings (references CHANGELOG.md on GitHub main branch, = this file)
 
 ### Changed
 - up tp 4 custom tabs rather than 3